+++ app/page.tsx
@@ -1,4 +1,4 @@
-import About from './components/About';
+import About from './_components/About';
 
 export default function Main() {
   return (
